listener: move active_udp_listener out of conn handler (#15349)
