+++ source/server/active_raw_udp_listener_config.cc
@@ -5,6 +5,7 @@
 
 #include "envoy/api/v2/listener/listener.pb.h"
 
+#include "server/active_udp_listener.h"
 #include "server/connection_handler_impl.h"
 #include "server/well_known_names.h"
